+++ assets/javascripts/kitten/components/form/input/text-input-with-button/index.js
@@ -6,6 +6,7 @@ import PropTypes from 'prop-types'
 import TYPOGRAPHY from '../../../../constants/typography-config'
 import { TextInput } from '../../../form/input/text-input'
 import { pxToRem, stepToRem } from '../../../../helpers/utils/typography'
+import { checkDeprecatedSizes } from '../../../../helpers/utils/deprecated'
 import { Button } from '../../../action/button'
 import { ScreenConfig } from '../../../../constants/screen-config'
 
@@ -61,15 +62,18 @@ const StyledTextInputWithButton = styled.div`
     }
   }
 
-  &.k-Form-TextInputWithButton--tiny {
+  &.k-Form-TextInputWithButton--tiny,
+  &.k-Form-TextInputWithButton--small {
     --text-input-size: ${pxToRem(40)};
   }
 
-  &.k-Form-TextInputWithButton--regular {
+  &.k-Form-TextInputWithButton--regular,
+  &.k-Form-TextInputWithButton--medium {
     --text-input-size: ${pxToRem(50)};
   }
 
-  &.k-Form-TextInputWithButton--big {
+  &.k-Form-TextInputWithButton--big,
+  &.k-Form-TextInputWithButton--large {
     --text-input-size: ${pxToRem(60)};
 
     @media (min-width: ${ScreenConfig.M.min}px) {
@@ -124,6 +128,8 @@ export const TextInputWithButton = React.forwardRef(
     },
     ref,
   ) => {
+    checkDeprecatedSizes(size)
+
     const ButtonComponent = inset ? 'button' : Button
 
     return (
@@ -176,7 +182,7 @@ TextInputWithButton.propTypes = {
   valid: PropTypes.bool,
   error: PropTypes.bool,
   disabled: PropTypes.bool,
-  size: PropTypes.string,
+  size: PropTypes.oneOf(['small', 'medium', 'large', 'huge', 'giant']),
   rounded: PropTypes.bool,
   inset: PropTypes.bool,
   modifier: PropTypes.string,
@@ -190,7 +196,7 @@ TextInputWithButton.defaultProps = {
   valid: false,
   error: false,
   disabled: false,
-  size: 'regular',
+  size: 'medium',
   buttonValue: 'Button',
   buttonProps: {
     disabled: false,