+++ assets/javascripts/kitten/components/information/boxes/doc-link-box/index.js
@@ -10,7 +10,7 @@ export const DocLinkBox = ({ isExternal, href, title, text, ...props }) => {
         <LinkBoxIllustration className="k-LinkBox__icon--svg" />
       </LinkBox.Icon>
       <LinkBox.Text>
-        <Text tag="span" weight="regular" size="tiny" lineHeight="normal">
+        <Text tag="span" weight="regular" size="small" lineHeight="normal">
           {title}
         </Text>
         <Text tag="span" weight="light" size="micro" lineHeight="normal">